added linint2 files to setup.py extensions, and refactored linint2_points names to be consistant.
