+++ src/geocat/ncomp/_linint2.pyx
